I moved the deleteFiles and deleteActiveTorrent methods from ActiveTorrents -->> ActiveTorrent class. Meanwhile, because I'm not using MongoDB, I must keep deleteActiveTorrent also inside ActiveTorrents because the list of ActiveTorrent are there (inside ActiveTorrents class) and I must access it to remove the ActiveTorrent object.
